+++ Plugins/Util/Util.cpp
@@ -402,6 +402,23 @@ ArgumentStack Util::AddScript(ArgumentStack&& args)
       ASSERT_OR_THROW(!scriptData.empty());
     const auto wrapIntoMain = Services::Events::ExtractArgument<int32_t>(args);
 
+    std::string alias;
+    try
+    {
+        alias = Services::Events::ExtractArgument<std::string>(args);
+    }
+    catch (const std::runtime_error& e)
+    {
+        LOG_WARNING("NWNX_Util_AddScript() called without alias parameter, please update nwnx_util.nss");
+        alias = "NWNX";
+    }
+
+    if (!Utils::IsValidCustomResourceDirectoryAlias(alias))
+    {
+        LOG_WARNING("NWNX_Util_AddScript() called with an invalid alias: %s, defaulting to 'NWNX'", alias);
+        alias = "NWNX";
+    }
+
     if (!m_scriptCompiler)
     {
         m_scriptCompiler = std::make_unique<CScriptCompiler>();
@@ -411,8 +428,9 @@ ArgumentStack Util::AddScript(ArgumentStack&& args)
         m_scriptCompiler->SetOptimizeBinaryCodeLength(true);
         m_scriptCompiler->SetCompileConditionalOrMain(true);
         m_scriptCompiler->SetIdentifierSpecification("nwscript");
-        m_scriptCompiler->SetOutputAlias("NWNX");
     }
+
+    m_scriptCompiler->SetOutputAlias(alias);
 
     if (m_scriptCompiler->CompileScriptChunk(scriptData.c_str(), wrapIntoMain != 0) == 0)
     {
@@ -460,7 +478,24 @@ ArgumentStack Util::AddNSSFile(ArgumentStack&& args)
       ASSERT_OR_THROW(fileName.size() <= 16);
     const auto contents = Services::Events::ExtractArgument<std::string>(args);
 
-    auto file = CExoFile(("NWNX:" + fileName).c_str(), Constants::ResRefType::NSS, "w");
+    std::string alias;
+    try
+    {
+        alias = Services::Events::ExtractArgument<std::string>(args);
+    }
+    catch (const std::runtime_error& e)
+    {
+        LOG_WARNING("NWNX_Util_AddNSSFile() called without alias parameter, please update nwnx_util.nss");
+        alias = "NWNX";
+    }
+
+    if (!Utils::IsValidCustomResourceDirectoryAlias(alias))
+    {
+        LOG_WARNING("NWNX_Util_AddNSSFile() called with an invalid alias: %s, defaulting to 'NWNX'", alias);
+        alias = "NWNX";
+    }
+
+    auto file = CExoFile((alias + ":" + fileName).c_str(), Constants::ResRefType::NSS, "w");
 
     if (file.FileOpened())
     {
@@ -482,7 +517,24 @@ ArgumentStack Util::RemoveNWNXResourceFile(ArgumentStack&& args)
       ASSERT_OR_THROW(fileName.size() <= 16);
     const auto type = Services::Events::ExtractArgument<int32_t>(args);
 
-    CExoString exoFileName = ("NWNX:" + fileName).c_str();
+    std::string alias;
+    try
+    {
+        alias = Services::Events::ExtractArgument<std::string>(args);
+    }
+    catch (const std::runtime_error& e)
+    {
+        LOG_WARNING("NWNX_Util_RemoveNWNXResourceFile() called without alias parameter, please update nwnx_util.nss");
+        alias = "NWNX";
+    }
+
+    if (!Utils::IsValidCustomResourceDirectoryAlias(alias))
+    {
+        LOG_WARNING("NWNX_Util_RemoveNWNXResourceFile() called with an invalid alias: %s, defaulting to 'NWNX'", alias);
+        alias = "NWNX";
+    }
+
+    CExoString exoFileName = (alias + ":" + fileName).c_str();
 
     retVal = Globals::ExoResMan()->RemoveFile(exoFileName, type);