+++ CHANGELOG.md
@@ -29,9 +29,12 @@ The following plugins were added:
 - Util: GetScriptParamIsSet()
 
 ### Changed
+- Area: ExportGIT() now supports valid custom resource directory aliases. 
 - Events: `NWNX_ON_DM_SPAWN_OBJECT_*` now provides the resref as event data.
 - Events: `NWNX_ON_STORE_REQUEST_*_AFTER` now provides the result as event data.
+- Events: ResourceEvents now support valid custom resource directory aliases.
 - Util: added an optional appearance type parameter to CreateDoor()
+- Util: AddScript(), AddNSSFile() and RemoveNWNXResourceFile() now support valid custom resource directory aliases.
 - Visibility: added two new visibility types to always show an object regardless of range.
 
 ### Deprecated